plan/requirements.txt
paho-mqtt==2.0.0
python-dotenv==1.0.1
carla
asn1tools==0.166.0
networkx==2.6.3
scipy==1.7.3